feat: Added opportunity to add items to different shopping varts

+++ views/user-area/shoppingCartList.php
@@ -1,6 +1,7 @@
 <?php
 
 use Vestis\Database\Models\ShoppingCart;
+use Vestis\Service\AuthService;
 
 /** @var ShoppingCart[] $ownedShoppingCarts */
 
@@ -45,7 +46,7 @@ use Vestis\Database\Models\ShoppingCart;
                         <td>
                             <div class="button-row">
                                 <a href="/user-area/shoppingCart?accId=<?= $ownedShoppingCart->accId ?>&cartNumber=<?= $ownedShoppingCart->cartNumber ?>" class="btn btn-sm">Öffnen.</a>
-                                <?php if ($ownedShoppingCart->cartNumber !== 1): ?>
+                                <?php if ($ownedShoppingCart->cartNumber !== 1 && $ownedShoppingCart->accId === AuthService::$currentAccount->id): ?>
                                     <a href="/user-area/shoppingCarts/delete?cartNumber=<?= $ownedShoppingCart->cartNumber ?>" class="btn btn-sm danger">Delete.</a>
                                 <?php endif; ?>
                             </div>
